Move percent complete logger to it's own file.
